% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definitions_magick.R
\name{annotation_magick}
\alias{annotation_magick}
\title{Load image to a ggplot2 layer}
\usage{
annotation_magick(picPath, interpolate = FALSE)
}
\arguments{
\item{picPath}{Image path for the correct position, only one.}

\item{interpolate}{Passed on to layer params.}
}
\value{
A custom annotation_raster for ggplot.
}
\description{
Funcion basada en magick para abrir una foto y mostrarla en un ggplot2 layer.
}
