% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hu_moments_short.R
\name{append_hues}
\alias{append_hues}
\title{Append Hu moments columns to data on a cell_data object}
\usage{
append_hues(cell_data, image_bits, return_points = F)
}
\arguments{
\item{cell_data}{a cell_data object as loaded by rcell2::load_cell_data}

\item{image_bits}{an integer indicating the bit-depth on the TIFF images.}

\item{return_points}{if TRUE it will add a "masks" dataframe to the cell_data object, containing the mask coordinates.}
}
\description{
It is imperative that the TIFF images contain only cell boundaries, with pixel intensity values such that "CellID = (2^image_bits -1) - pixel_value"
}
\details{
This is automatically generated by the mask_mod branch of the CellID program found at: https://github.com/darksideoftheshmoo/cellID-linux/tree/mask_mod

To play with the XY coordinates of the TIFF masks you may want to set \code{return_points = T}.
Or if you want to skip calculation of the Hu moments, checkout \code{rcell2:::pic_df_from_tiff}.
}
