% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hu_moments_short.R
\name{append_hues2}
\alias{append_hues2}
\title{Generate Hu moments from CellID's mask TSV files}
\usage{
append_hues2(
  tsv_files_df,
  cell_data = NULL,
  parralellize = T,
  shape_pixtype = "b",
  shape_flagtype = 0,
  overwrite = F
)
}
\arguments{
\item{tsv_files_df}{A data.frame with paths to the TSV files with the cells' XY coordinates (requires "path" and "pos" columns). Use \code{tsv_paths_from_args()} for convenience.}

\item{cell_data}{Either NULL, a list, or cell.data object from rcell2::load_cell_data. If not NULL, the result is added to the list.}

\item{parralellize}{Enable cell-wise parallelization using parallel::parLapply}

\item{shape_pixtype}{Default "b" for Hu moments based on boundary points. A character vector containing any of c("b", "i").}

\item{shape_flagtype}{Default 0 for Hu moments based on flag value 0. Can be any of the integer flag values present in the \code{out_bf_fl_mapping} CellID files.}

\item{overwrite}{Default FALSE, overwrite Hu moments element in the input list if already present.}
}
\value{
The Hu moments dataframe usis assgned to the input list as an element named "Hu_moments".
}
\description{
The required TSV are generated by the mask_mod branch of the CellID program found at: https://github.com/darksideoftheshmoo/cellID-linux/tree/mask_mod
}
