% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definitions.R
\name{applyFilter}
\alias{applyFilter}
\title{Add TRUTH column to cdata}
\usage{
applyFilter(
  cdataDF,
  filterDF,
  cell_unique_id_field = "ucid",
  truth_column = "filter"
)
}
\description{
Add TRUTH column to cdata
}
