% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellid_wrapper.R
\name{arguments_to_images}
\alias{arguments_to_images}
\title{Make "images" dataframe from "arguments" dataframe}
\usage{
arguments_to_images(arguments)
}
\value{
A data.frame similar to \code{cell.load.alt()$images}.
}
\description{
Essentially a pivot_longer of the arguments.
}
