% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definitions_shinycell_utils.R
\name{bind_filters}
\alias{bind_filters}
\title{Bind shinyCell polygon filters by variable pairs}
\usage{
bind_filters(saved_data, print_plots = TRUE)
}
\arguments{
\item{saved_data}{The output of shinyCell.}
}
\value{
A list of polygons bound by variable, with names unique to variable pairs (by sort). "x" and "y" column names may be swapped.
}
\description{
Useful to check out what areas the filters are covering.
}
