% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k-means_clustering.R
\name{calculate_extra_morpho_vars}
\alias{calculate_extra_morpho_vars}
\title{calculate_extra_morpho_vars}
\usage{
calculate_extra_morpho_vars(x)
}
\arguments{
\item{x}{a cell.data data.frame}
}
\value{
a cell.data data.frame to which new variables have been added
}
\description{
Create custom fit variables
}
