% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculate_ntot}
\alias{calculate_ntot}
\title{calculate number of time frames a cell appears in}
\usage{
calculate_ntot(data)
}
\arguments{
\item{data}{data.frame}
}
\value{
data.frame with n.tot variable
}
\description{
In case a cell was filtered out, its n.tot will be 0.
}
