% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellid_wrapper.R
\name{cargar.out_all}
\alias{cargar.out_all}
\title{Una función para leer y combinar todos los archivos "out".}
\usage{
cargar.out_all(
  .carpeta,
  position.pattern = ".*Position(\\\\d+).*",
  out_file_pattern = "^out_all$",
  out_mapping_pattern = "^out_bf_fl_mapping$",
  fluorescence.pattern = ".*([A-Z])FP_Position.*"
)
}
\arguments{
\item{.carpeta}{path a la carpeta donde están os output de CellID}

\item{position.pattern}{Regex describing what the position string looks like (default 'Position\\d+')}

\item{fluorescence.pattern}{Regex describing what the fluorescence string looks like (default ".*([GCYRT])FP_Position.*")}

\item{.nombre.archivos}{paths de los "out_all"}

\item{.nombre.archivos.map}{paths de los "bf_fl_mapping"}
}
\value{
A list of two dataframes: `d` contains the actual output, and `out.map` contains image paths and metadata.
}
\description{
Una función para leer y combinar todos los archivos "out".
}
