% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellid_wrapper.R
\name{cell.images.BF}
\alias{cell.images.BF}
\title{Buscar BFs}
\usage{
cell.images.BF(path, BF.pattern = "BF_Position\\\\d+\\\\.tif$")
}
\arguments{
\item{path}{directory where images are stored, full path.}

\item{pattern}{BF pattern "BF_Position\\d+\\.tif$"}
}
\value{
A vector of directories.
}
\description{
Buscar BFs
}
