% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellid_wrapper.R
\name{cell.images.FP}
\alias{cell.images.FP}
\title{Buscar ?FPs}
\usage{
cell.images.FP(path, pattern = "FP_Position\\\\d*\\\\.tif$")
}
\arguments{
\item{path}{directory where images are stored, full path.}

\item{pattern}{BF pattern "FP_Position\\d*\\.tif$"}
}
\value{
A vector of directories.
}
\description{
Buscar ?FPs
}
