% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellid_wrapper.R
\name{cell}
\alias{cell}
\title{Filtrar cdata usando gráficos y dibujando regiones}
\usage{
cell(
  cell.args,
  path = "data/images/",
  position.pattern = "Position\\\\d+",
  cell.command = "~/Software/cellID-linux/cell",
  channels = c("b", "f"),
  no_cores = NULL
)
}
\arguments{
\item{cell.args}{An argument list, as built by cellArgs().}

\item{path}{directory where images are stored, full path.}

\item{position.pattern}{a regular expression that recognizes the position in the image name}

\item{cell.command}{the CellID command, either "cellBUILTIN" for the builtin binary, a path to it.}

\item{channels}{Default c("b", "f"), don't change, used to create temporal file lists for BF and ?FP.}
}
\value{
Nothing.
}
\description{
Filtrar cdata usando gráficos y dibujando regiones
}
