% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellid_wrapper.R
\name{cell2}
\alias{cell2}
\title{Function to run CellID}
\usage{
cell2(
  arguments,
  cell.command = NULL,
  no_cores = NULL,
  debug_flag = 0,
  dry = F,
  label_cells_in_bf = F
)
}
\arguments{
\item{arguments}{An argument data.frame, as built by rcell2::cellArgs2.}

\item{cell.command}{the CellID command, either "cellBUILTIN" for the builtin binary, a path to the binary executable (get if from https://github.com/naikymen/cellID-linux).}

\item{no_cores}{Position-wise parallelization,internally capped to number of positions in cell.args.}

\item{debug_flag}{Set to 0 to disable CellID printf messages.}

\item{dry}{Do everything without actually running CellID.}

\item{label_cells_in_bf}{Set to TRUE to enable lableing of cells with their CellID in the BF image.}
}
\value{
Nothing :) use rcell2::load_cell_data to get the results from the output at the images path
}
\description{
Function to run CellID
}
