% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellid_wrapper.R
\name{cellArgs}
\alias{cellArgs}
\title{Filtrar cdata usando gráficos y dibujando regiones}
\usage{
cellArgs(path, parameters)
}
\arguments{
\item{path}{directory where images are stored, full path.}

\item{parameters}{absolute path to the parameters file.}

\item{b}{list of arguments: paths to BF images}

\item{f}{list of arguments: paths to ?FP images}

\item{o}{list of arguments: paths to output directories}

\item{channels}{Default c("b", "f"), don't change, used to create temporal file lists for BF and ?FP.}
}
\value{
Nothing.
}
\description{
Filtrar cdata usando gráficos y dibujando regiones
}
