% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellid_wrapper.R
\name{cellArgs2}
\alias{cellArgs2}
\title{Obtener argumentos para CellID}
\usage{
cellArgs2(
  path,
  parameters,
  BF.pattern = "^BF",
  file.pattern = "^(BF|[A-Z]FP)_Position(\\\\d+)_time(\\\\d+).tif$",
  file.pattern.groups.order = c("ch", "pos", "t.frame"),
  out.dir = "out",
  tiff.ext = "tif$"
)
}
\arguments{
\item{path}{directory where images are stored, full path.}

\item{parameters}{path to the parameters file or a data.frame with "pos" (position number) and "parameter" (path) columns.}

\item{BF.pattern}{regex pattern to BF images.}

\item{file.pattern}{regex pattern for all tif files, with one group for each of: c("ch", "pos", "t.frame")}

\item{file.pattern.groups.order}{a character vector of components c("ch", "pos", "t.frame") with order corresponding to the order of groups in file.pattern}

\item{out.dir}{name for output directories paths "out"}

\item{tiff.ext}{regex pattern for the tif file extension}
}
\value{
a data.frame with all the information needed to run CellID
}
\description{
Obtener argumentos para CellID
}
