% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellid_wrapper.R
\name{cellid}
\alias{cellid}
\title{Correr cellid desde C}
\usage{
cellid(args)
}
\arguments{
\item{args}{el comando de cellid entero, tal como se ejecutaria en bash "cell -p ..."}
}
\value{
Nada, el output está en los directorios.
}
\description{
Correr cellid desde C
}
