% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellid_wrapper.R
\name{cellid_parameter_descriptions}
\alias{cellid_parameter_descriptions}
\title{Cell-ID parameter descriptions}
\usage{
cellid_parameter_descriptions(list_format = T)
}
\arguments{
\item{list_format}{If TRUE then format the dataframe into a named list}
}
\description{
Cell-ID parameter descriptions
}
