% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definitions_magick.R
\name{cellSpreadPlot}
\alias{cellSpreadPlot}
\title{Plot for 2D binning of data and tiling of cell pictures}
\usage{
cellSpreadPlot(
  cdata,
  paths,
  ch = "BF.out",
  boxSize = 80,
  xvar = "a.tot",
  yvar = "fft.stat",
  overlay_points = F,
  underlay_points = F,
  draw_contour_breaks = NULL,
  x.cuts = 7,
  y.cuts = 7,
  ...
)
}
\arguments{
\item{cdata}{A Rcell data.frame (not the object).}

\item{paths}{A paths dataframe with file path, t.frame, position and channel information of each picture.}

\item{ch}{Name of the CellID channel (BF, BF.out, RFP, etc.). "BF.out" by default.}

\item{boxSize}{Size of the box containing the individual cell images. 50 by default.}

\item{xvar}{Strings indicating names for the variables to plot in the horizontal (x) and vertical (y) axis.}

\item{yvar}{Strings indicating names for the variables to plot in the horizontal (x) and vertical (y) axis.}

\item{overlay_points}{Overlay data points to the image plot.}

\item{underlay_points}{Underlay data points to the image plot.}

\item{draw_contour_breaks}{Overlay a ggplot2::stat_density2d layer. If TRUE, use the default breaks. Otherwise NULL for none, or a numeric vector for the density breaks.}

\item{x.cuts}{Integers indicating the number of cuts for each variable.}

\item{y.cuts}{Integers indicating the number of cuts for each variable.}

\item{...}{Arguments passed on to magickCell.}
}
\description{
Plot for 2D binning of data and tiling of cell pictures
}
