% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definitions_magick.R
\name{cellStrips}
\alias{cellStrips}
\title{Wraps cellMagick to make strips, optionally cutting them.}
\usage{
cellStrips(
  cdata,
  paths,
  n_ucids = NULL,
  cut_breaks = 1,
  split_col = "ucid",
  ch = c("BF.out", "YFP.out"),
  sortVar = "t.frame",
  ...
)
}
\arguments{
\item{cdata}{A Rcell data.frame (not the object).}

\item{paths}{A paths dataframe with file path, t.frame, position and channel information of each picture.}

\item{n_ucids}{will select the first `n_ucids`}

\item{...}{
  Arguments passed on to \code{\link[=cellStrip]{cellStrip}}
  \describe{
    \item{\code{stack_time_horizontally}}{Time increases from left to right (TRUE) or from up to down (FALSE).}
    \item{\code{channels}}{Name of the CellID channel (BF, BF.out, RFP, etc.). "BF.out" by default.}
    \item{\code{equalize_images}}{Use magick's function to "equalize" the image when TRUE (FALSE by default). Can be a logical vector, each value applied separately to each channel (recycled to the length of \code{ch}).}
    \item{\code{normalize_images}}{Use magick's function to "normalize" the image when TRUE (FALSE by default). Can be a logical vector, each value applied separately to each channel (recycled to the length of \code{ch}).}
  }}

\item{cut_strips}{Use `cut` to split the image series (by index; preserves sortVar order).}
}
\description{
`cdata` is split by `split_col` and then images are generated.
}
\details{
`images` are split with `cut`, useful wen strips are too long.
}
