% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cimage.R
\name{cimage.cell.image}
\alias{cimage.cell.image}
\title{cimage.cell.image}
\usage{
\method{cimage}{cell.image}(
  X,
  formula = NULL,
  subset = NULL,
  facets = NULL,
  scales = "fixed",
  allow.expressions = FALSE,
  nx = NULL,
  ny = NULL,
  facets.nx = NULL,
  facets.ny = NULL,
  bg.col = "white",
  border = 1,
  facets.border = 1,
  rev.y = TRUE,
  font.size = 14,
  font.col = "black",
  display = interactive(),
  ...
)
}
\arguments{
\item{X}{cell.data or cell.image object to plot}

\item{formula}{formula of the form 'var1+var2~var3' specifying how the images are to be ordered. See details.}

\item{subset}{logical expression indicating elements or rows to keep. Don't specify channel here}

\item{facets}{formula of the form 'var1+var2~var3' specifying how to facet the plot. See details.}

\item{scales}{either 'none', 'fixed' or 'free' axis for each facet}

\item{allow.expressions}{allow expressions in formulas, set to TRUE when called from cimage.cell.data}

\item{nx}{number of columns of images within each facet. Used with \code{formula} '~var1' or 'var1~.'}

\item{ny}{number of rows of images within each facet. Used with \code{formulas} '~var1' or 'var1~.'}

\item{facets.nx}{number of columns of facets. Used with \code{facets} '~var1' or 'var1~.'}

\item{facets.ny}{number of rows of facets. Used with \code{facets} '~var1' or 'var1~.'}

\item{bg.col}{The background color of the plot}

\item{border}{the width in pixels of the border between images}

\item{facets.border}{the width in pixels of the border between facets}

\item{rev.y}{boolean indicating if the y axis should be reversed}

\item{font.size}{The size of the font to use, in pixels}

\item{font.col}{The color of the font to use}

\item{display}{boolean indicating if the created image should be displayed}

\item{...}{further arguments for methods. \code{cimage} calls \code{\link{get_cell_image}},
so all the arguments of this function are available.}
}
\value{

}
\description{
returns cropped cell images arranged in some way
}
