% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cimage.R
\name{cimage}
\alias{cimage}
\title{cimage}
\usage{
cimage(X, ...)
}
\arguments{
\item{X}{cell.data object}

\item{...}{arguments for \code{\link{cimage.cell.image}} and \code{\link{cimage.cell.data}}}
}
\value{

}
\description{
returns a cell.image object
}
