% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magrittr_compatibility.R
\name{d}
\alias{d}
\title{d}
\usage{
d(cell.data)
}
\arguments{
\item{cell.data}{cell.data object}
}
\value{
a tibble
}
\description{
convenience function to use when piping.
}
\details{
d(X) is equivalent to X$data
}
