% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_cell_data.R
\name{.format.sequence}
\alias{.format.sequence}
\title{Format sequence of numbers}
\usage{
.format.sequence(pos)
}
\arguments{
\item{pos}{last number in sequence to shorten}
}
\value{
string of shortened sequence
}
\description{
formats sequence of numbers in an short expresion  eg: 1-10, 12-15
}
\keyword{internal}
