% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_cell_data.R
\name{.mk_flag_table}
\alias{.mk_flag_table}
\title{Make Flag Table}
\usage{
.mk_flag_table(bf.fl.mapping, pos = NULL)
}
\arguments{
\item{bf.fl.mapping}{data.frame of the bf.fl.mapping output file}

\item{pos}{integer, corresponds to position}
}
\value{
a data.frame containing the bf.fl.mapping of a single position
}
\description{
Generates a table mapping a channel name (3 first characters of the image file)
to a flag number for a given bf.fl.mapping data.frame (read from an
output_bf_fl_mapping file)
}
