% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_cell_data.R
\name{.parse_load_vars}
\alias{.parse_load_vars}
\title{Parse Names of Variables to Load}
\usage{
.parse_load_vars(load.vars, vars.all = NULL)
}
\arguments{
\item{load.vars, }{pattern of variable names in code}

\item{vars.all}{}
}
\value{
character vector containing variable names
}
\description{
Parses the input of load.vars and reurns a vector with the elements to be loaded.
Possible values for load.vars are 'all', 'fl' or 'fluorescence', 'bg' or 'background', 'calc',
'morph' or 'morphological', 'vac' or 'vacuole', 'nucl' or 'nuclear', 'disc'.
The group of variables can be specified in either a positive form (i.e. '+fl+bg+morph')
or in a negative form (i.e. '-nucl-vac'). Combination of positive and negative form is not allowed.
}
