% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k-means_functions.R
\name{get.fit.variables}
\alias{get.fit.variables}
\title{get.fit.variables}
\usage{
get.fit.variables(x, vars = "m")
}
\arguments{
\item{x}{}

\item{vars}{"m" for morph.vars, "f" for fluor.vars}
}
\value{
data.frame with fit.variables ("extr.vars") only
}
\description{
Extract columns for classification
}
