% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cimage.R
\name{get_cell_image.cell.data}
\alias{get_cell_image.cell.data}
\title{get_cell_image}
\usage{
\method{get_cell_image}{cell.data}(
  X,
  subset = NULL,
  channel.subset = NULL,
  channel = NULL,
  time.course = TRUE,
  group = NULL,
  na.rm = TRUE,
  N = 7,
  select = NULL,
  exclude = NULL,
  qc.filter = TRUE,
  box.size = 20,
  ...
)
}
\arguments{
\item{X}{cell.data object or data.frame that specifies the images}

\item{subset}{logical expression indicating elements or rows to keep. Don't specify channel here.}

\item{channel.subset}{logical expression to specify which image to retrieve with channel and t.frame variables.}

\item{channel}{character vector of channels to retrieve. If specified, defines the order of the channels.}

\item{time.course}{boolean indicating if the desired image montage is a time course (i.e. several images for the same cell)}

\item{group}{character vector or quoted names of variables who's interaction define the groups from which select \code{N} random cells.}

\item{na.rm}{boolean indicating if NAs should be removed.}

\item{N}{Number of random cells to select from each group. If NULL all cells are selected}

\item{select}{character vector defining variables names to be included in the returned cell.image object}

\item{exclude}{character vector defining variables names to be excluded from the returned cell.image object}

\item{qc.filter}{a boolean value indicating if the quality control filter should be applied over the data}

\item{box.size}{size in pixels of the image containing the cells. This specifies the 'radius', i.e. the image will be a square of length 2*box.size+1}

\item{...}{further arguments for methods}
}
\value{

}
\description{
Retrieves the images from single cells in an cell.image object
}
