% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cimage.R
\name{get_cell_image.data.frame}
\alias{get_cell_image.data.frame}
\title{get_cell_image.data.frame}
\usage{
\method{get_cell_image}{data.frame}(X, box.size = 20, contained.box = FALSE, bg.col = 0, ...)
}
\arguments{
\item{X}{cell.data object or data.frame that specifies the images}

\item{box.size}{size in pixels of the image containing the cells. This specifies the 'radius', i.e. the image will be a square of length 2*box.size+1}

\item{contained.box}{boolean indicating if the XY position of the box should be corrected to be contained in the original image. Relevant for cells close to the image border. If FALSE the part of the box outside the original image will be filled with \code{bg.col}}

\item{bg.col}{color to be used for the background of the images}

\item{...}{further arguments for methods}
}
\value{

}
\description{
get_cell_image.data.frame
}
