% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cimage.R
\name{get_cell_image}
\alias{get_cell_image}
\title{Retrieves the images from single cells in an cell.data or cell.image (dataframe) object}
\usage{
get_cell_image(X, ...)
}
\arguments{
\item{X}{cell.data object or data.frame that specifies the images}

\item{...}{further arguments for methods}
}
\description{
Retrieves the images from single cells in an cell.data or cell.image (dataframe) object
}
