% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_tiff_tags.R
\name{get_stage_xy_from_file}
\alias{get_stage_xy_from_file}
\title{Get stage XY coordinates from tiff file's description tag}
\usage{
get_stage_xy_from_file(tiff.path, frames = 1)
}
\arguments{
\item{tiff.path}{A path to the tiff file, passed to \code{\link[ijtiff]{read_tags}}.}

\item{frames}{The frame number to process, passed to \code{\link[ijtiff]{read_tags}}.}
}
\value{
A named vector, with "x" and "y" coordinates.
}
\description{
The description is in the tiff file's metadata, tipically written by Metamorph.
}
\examples{

# Get the "images" data.frame from CellID
cell.data <- get_cell_data("path/to/output/")
images <- cell.data$images

# Generate a list with file paths split and named by "pos"
tags.list <- images \%>\% filter(t.frame==0, channel=="BF") \%>\% 
  {split(.$file, .$pos)}
  
# Otherwise, a simple list of file paths ordered by "pos" suffice.
# tags.list <- list(file.paths)  # NOT RUN

# Get stage positions
tags.stage <- tags.list \%>\% 
  lapply(get_stage_xy_from_file) \%>\% bind_rows(.id="pos") \%>\% 
  mutate(pos = as.integer(pos))

tags.stage

}
