% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_tiff_tags.R
\name{get_tiff_description}
\alias{get_tiff_description}
\title{Get Metamorph image XML description}
\usage{
get_tiff_description(tiff.path, frames = 1)
}
\arguments{
\item{tiff.path}{A path to the tiff file, passed to ijtiff::read_tags().}

\item{frames}{The frame number to process, passed to ijtiff::read_tags().}
}
\description{
Get Metamorph image XML description
}
