% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definitions_magick.R
\name{getCellGeom}
\alias{getCellGeom}
\title{Geometria para agarrar un cuadradito de una imagen con magick por coordenada del centro}
\usage{
getCellGeom(xpos, ypos, boxSize = 50)
}
\arguments{
\item{xpos}{x pixel coordinate.}

\item{ypos}{y pixel coordinate.}

\item{boxSize}{side length of the final square}
}
\value{
magick::geometry_area configured for cellMagick
}
\description{
https://ropensci.org/blog/2017/08/15/magick-10/
}
\details{
https://livefreeordichotomize.com/2017/07/18/the-making-of-we-r-ladies/
}
