% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definitions.R
\name{getPositions}
\alias{getPositions}
\title{Get a list of numbers from the "position" input string}
\usage{
getPositions(input, numPos)
}
\description{
Get a list of numbers from the "position" input string
}
