% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw.R
\name{helloWorld}
\alias{helloWorld}
\title{Say hi!}
\usage{
helloWorld(dummy)
}
\arguments{
\item{dummy}{A dummy parameter that does nothing. It's super effective!}
}
\description{
Say hi!
}
\examples{
helloWorld(":)")
}
\seealso{
\link{print}
}
