% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definitions_hex.R
\name{hexPlotDf}
\alias{hexPlotDf}
\title{Funcion para armar un dataframe para hacer plots tipo HEX}
\usage{
hexPlotDf(
  cdata,
  facetVars = c("pos", "treatment"),
  varx = "a.tot",
  vary = "el.p"
)
}
\arguments{
\item{cdata}{A "cdata" Rcell data.frame (not the object).}

\item{facetVars}{are "pdata" variables}

\item{varx}{the x-axis "pdata" variable}

\item{vary}{the y-axis "pdata" variable}
}
\value{
Lots of stuff.
}
\description{
Funcion para armar un dataframe para hacer plots tipo HEX
}
