% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definitions.R
\name{hover_closest}
\alias{hover_closest}
\title{Find cell closest to hover point in shiny}
\usage{
hover_closest(ui_input, cdata)
}
\arguments{
\item{ui_input}{the "input" object from whiny server}

\item{cdata}{the dataframe to filter}
}
\description{
Find cell closest to hover point in shiny
}
