% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hu_moments_short.R
\name{hu.moments}
\alias{hu.moments}
\title{Hu moments}
\usage{
hu.moments(xy)
}
\arguments{
\item{xy}{Matriz de dos columnas (dim1 y dim2), con las coordenadas XY del boundary binario. No es importante el orden de los puntos.}
}
\value{
Los 7 Hu moments + 1 extra (?) en un named vector
}
\description{
This function is CellID agnostic.
}
