% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hu_moments_short.R
\name{hues_from_tsv2}
\alias{hues_from_tsv2}
\title{Generate Hu moments from XY coordinates TSV file}
\usage{
hues_from_tsv2(
  masks_tsv_path,
  .parallel = F,
  shape_pixtype = NULL,
  shape_flagtype = NULL
)
}
\arguments{
\item{masks_tsv_path}{A path to the TSV file holding XY coordinates, from CellID's output with "-t" option.}

\item{.parallel}{Enable cell-wise parallelization using parallel::parLapply}

\item{shape_pixtype}{Default "b" for Hu moments based on boundary points. A character vector containing any of c("b", "i").}

\item{shape_flagtype}{Default 0 for Hu moments based on flag value 0. Can be any of the integer flag values present in the \code{out_bf_fl_mapping} CellID files.}
}
\description{
Cells in the dataframe are split by cellID, t.frame, flag, and pixtype by default.
}
\details{
Optional filtering by shape type (boundary or interior) and CellID flag number.

Optional cell-wise parallelization.
}
