% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hu_moments_short.R
\name{hues_from_tsv_files2}
\alias{hues_from_tsv_files2}
\title{Generate Hu moments data frame from one or more TSV files}
\usage{
hues_from_tsv_files2(
  tsv_files_df,
  return_points = F,
  parralellize = T,
  shape_pixtype = NULL,
  shape_flagtype = NULL,
  cdata_subset = NULL
)
}
\arguments{
\item{return_points}{if TRUE it will add a "masks" dataframe to the cell_data object, containing the mask coordinates.}

\item{parralellize}{Enable cell-wise parallelization using parallel::parLapply}

\item{shape_pixtype}{Default "b" for Hu moments based on boundary points. A character vector containing any of c("b", "i").}

\item{shape_flagtype}{Default 0 for Hu moments based on flag value 0. Can be any of the integer flag values present in the \code{out_bf_fl_mapping} CellID files.}

\item{cdata_subset}{Subset of cdata with unique rows, and only pos, t.frame and cellID columns.}

\item{tsv_files}{paths to the XY coordinates TSV files from CellID mask_mod "-t" option.}
}
\description{
The required TSV are generated by the mask_mod branch of the CellID program found at: https://github.com/darksideoftheshmoo/cellID-linux/tree/mask_mod
}
\details{
To play with the XY coordinates of the TIFF masks you may want to set \code{return_points = T}.
}
