% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definitions_magick.R
\name{image_border_one}
\alias{image_border_one}
\title{Add a border to one side of the image}
\usage{
image_border_one(image, geometry = "0x15", color = "white")
}
\arguments{
\item{image}{magick image.}

\item{geometry}{magick geometry string.}

\item{color}{color string for the border background.}
}
\description{
magick's image_border adds borders symmetrically to all sides. image_border_one is good for padding images to add captions later.
}
