% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagej_macros.R
\name{imagej.fft.filter}
\alias{imagej.fft.filter}
\title{Run ImageJ FFT filter macro from R}
\usage{
imagej.fft.filter(
  pic.path,
  script.path = system.file("inst/imagej_macros/FFT_filter_on_BFs_R.txt", package =
    "rcell2")
)
}
\arguments{
\item{pic.path}{Path to the directory containing the image files.}

\item{script.path}{Path to the ImageJ macro. Defaults to built-in macro.}
}
\description{
Passes an ImageJ FFT filter on files matching \code{BF.*.tif} in the target directory.
}
\details{
The modified images are saved to a new subdirectory with default name: "filtered/".
}
