% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagej_macros.R
\name{imagej.macro.run}
\alias{imagej.macro.run}
\title{Run headless ImageJ Macro file}
\usage{
imagej.macro.run(
  script.path,
  ...,
  imagej.path = "~/Software/Fiji.app/ImageJ-linux64",
  wait = T,
  headless = T
)
}
\arguments{
\item{script.path}{Path to the ImageJ macro. Defaults to built-in macro.}

\item{imagej.path}{Path to the ImageJ binary (a path to "ImageJ-linux64" or equivalent).}

\item{wait}{a logical (not \code{NA}) indicating whether the \R
    interpreter should wait for the command to finish, or run it
    asynchronously.  This will be ignored (and the interpreter will
    always wait) if \code{intern = TRUE}. When running the command
    asynchronously, no output will be displayed on the \code{Rgui}
    console in Windows (it will be dropped, instead).}
}
\description{
Run headless ImageJ Macro file
}
