% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_cell_data.R
\name{is.cell.data}
\alias{is.cell.data}
\title{is cell data}
\usage{
is.cell.data(X)
}
\arguments{
\item{X}{an object}
}
\value{
boolean indicating if X is a cell.data object
}
\description{
checks if an objects is a cell.data object
}
