% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k-means_functions.R
\name{k.means.classification}
\alias{k.means.classification}
\title{k.means.classification}
\usage{
k.means.classification(x, k = 10, max.iterations = 100, vars = NULL)
}
\arguments{
\item{x}{possibly the cell.data object}

\item{k}{}

\item{vars}{}
}
\value{
the data with a classification column added as "k.data"
}
\description{
Perform k-means classification to Cell-ID data. Appends a 'k' column indicating the assigned class, and a 'k.dist' column indicating the calculated Euclidean distance to the nearest centroid.
}
