% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k-means_functions.R
\name{kmeans_clustering}
\alias{kmeans_clustering}
\title{kmeans_clustering}
\usage{
kmeans_clustering(
  x,
  k = 10,
  max_iter = 100,
  resume = FALSE,
  tag_col = "k",
  var_cats = "m",
  custom_vars = NULL
)
}
\arguments{
\item{x}{cell.data object}

\item{k}{non-negative integer. The desired number of clusters.}

\item{max_iter}{non-negative integer. The maximal number of iterations allowed, if the algorithm has not yet converged.}

\item{resume}{logical. If TRUE the algorithm resumes clustering from previously assigned clusters.}

\item{tag_col}{optional string defining the column containing pre-defined tags. Only utilized when 'resume' is set to TRUE. Defaults to 'k' if not defined.}

\item{var_cats}{optional character vector specifying whether morphological ('morpho') and/or fluorescence ('fluor') variables should be considered for clustering.}

\item{custom_vars}{optional character vector, specifies variables to be considered for clustering. These variables are}

\item{centroids}{optional integer vector of length 'k' defining the rows in 'x$data' to be chosen as starting centroids. If not provided, these are selected randomly.}
}
\value{
a cell.data object with columns 'k' and 'k.dist' appended to the data.frame, indicating the assigned class and Euclidean distance to the assigned centroid, respectively.
}
\description{
Perform k-means clustering on Cell-ID data.
}
