% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_cell_data.R
\name{load_cell_data_checkid}
\alias{load_cell_data_checkid}
\title{Check UCID-T.FRAME unicity}
\usage{
load_cell_data_checkid(cdata, id_cols = c("ucid", "t.frame"))
}
\arguments{
\item{cdata}{The "data" item of the list produced by \code{load_cell_data} or \code{get_cell_data}.}

\item{id_cols}{Character vector with the column names of cdata that must identify rows uniquely. Defaults to the usual ucid-t.frame combination.}
}
\value{
Invisibly returns a named list with the test result ("all_unique": boolean) and the summarized dataframe ("cdata_summary": data.frame) used for checking. Any value of n_rows > 1 is problematic.
}
\description{
The ucid-t.frame combination should be unique. Else there might be a problem (e.g. redundant matches in pdata).
}
