% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hu_moments_short.R
\name{load_tiff_masks}
\alias{load_tiff_masks}
\title{Load cell masks from Bf.out TIFF files}
\usage{
load_tiff_masks(images, image_bits, cell_id_offset = -1, return_points = F)
}
\arguments{
\item{images}{An images dataframe, from  \code{rcell2::load_cell_data} or \code{rcell2.rcellid::cell.load.alt}.}

\item{image_bits}{an integer indicating the bit-depth on the TIFF images, such that maximum intensity is \code{image_bits^2 -1} and minimum is zero. Tipically 8 or 16.}

\item{cell_id_offset}{the offset respect to maximum pixel intensity, such that \code{cellID = maximum_intensity - boundary_intensity + cell_id_offset}.}

\item{return_points}{if TRUE it will add a "masks" dataframe to the cell_data object, containing the mask coordinates.}
}
\description{
It is imperative that the TIFF images contain only cell boundaries, with pixel intensity values such that \code{CellID = (2^image_bits -1) - pixel_value}.
}
\details{
Such TIFF images are automatically generated by the mask_mod branch of the CellID program found at: https://github.com/darksideoftheshmoo/cellID-linux/tree/mask_mod
}
