% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hu_moments_short.R
\name{load_tsv_masks}
\alias{load_tsv_masks}
\title{Generate a data.frame with XY mask data from CellID's TSV files}
\usage{
load_tsv_masks(
  dir_path,
  tsv_pattern = "^out_all_masks.tsv$",
  position_pattern = ".*Position(\\\\d+)$",
  ...
)
}
\arguments{
\item{dir_path}{Path to the directory containing CellID's outputs (tipically the images' directory). CellID must have been run with TSV output enabled.}

\item{tsv_pattern}{A regex pattern matching the names of the TSV files.}

\item{position_pattern}{A regex pattern with one group for the integer position number, extracted from the directory name holding the TSV file.}

\item{...}{
  Arguments passed on to \code{\link[=read_coords_tsv]{read_coords_tsv}}
  \describe{
    \item{\code{masks_tsv_path}}{A path to the TSV file holding XY coordinates, from CellID's output with "-t" option.}
    \item{\code{shape_pixtype}}{Default "b" for Hu moments based on boundary points. A character vector containing any of c("b", "i").}
    \item{\code{shape_flagtype}}{Default 0 for Hu moments based on flag value 0. Can be any of the integer flag values present in the \code{out_bf_fl_mapping} CellID files.}
  }}
}
\value{
A data.frame with XY
}
\description{
Generate a data.frame with XY mask data from CellID's TSV files
}
\details{
Do not specify masks_tsv_path, it is generated by tsv_paths_from_dir automatically.
}
