% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magick_helpers.R
\name{magickPaths}
\alias{magickPaths}
\title{Extaer paths para cellMagick del objeto cell.data}
\usage{
magickPaths(cell.data)
}
\arguments{
\item{cell.data}{directory where images are stored, full path.}
}
\value{
A dataframe with paths.
}
\description{
Extaer paths para cellMagick del objeto cell.data
}
