% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k-means_functions.R
\name{make.fit.variables}
\alias{make.fit.variables}
\title{make.fit.variables}
\usage{
make.fit.variables(x)
}
\arguments{
\item{X}{a cell.data data.frame to which fir.vars are added}
}
\value{
Normalized X
}
\description{
Create custom fit variables
}
