% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cell_mask_data.R
\name{mask_df_from_tiff}
\alias{mask_df_from_tiff}
\title{Read cell masks from BF.out}
\usage{
mask_df_from_tiff(tiff_path, image_bits, cell_id_offset = -1)
}
\description{
Generate a dataframe with boundary coordinates from mask tiff file.
}
\details{
This function is alternative to Andy's much smarter \code{read_tiff_masks} function.
}
\keyword{internal}
