% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definitions_hex.R
\name{myhexbin}
\alias{myhexbin}
\title{Función hexbin}
\usage{
myhexbin(bindata, varx, vary, xbins = 25)
}
\description{
Esencialmente hexbin() pero un poco más a mi gusto.
}
\details{
La función puede fallar de forma medio oscura cuando el input tiene problemas
(i.e. si está vacío, o si es todo NAs, etc.)
}
\keyword{internal}
