% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definitions.R
\name{numbers_to_intervals}
\alias{numbers_to_intervals}
\title{convert sequence of numbers to ranges}
\usage{
numbers_to_intervals(numbers = c(1:10, 20:30))
}
\description{
convert sequence of numbers to ranges
}
