% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellid_wrapper.R
\name{parameters.default}
\alias{parameters.default}
\title{Default parameters list for Cell-ID}
\usage{
parameters.default(
  max_split_over_minor = 0.5,
  max_dist_over_waist = 8,
  max_pixels_per_cell = 2000,
  min_pixels_per_cell = 75,
  background_reject_factor = 0.75,
  tracking_comparison = 0.2,
  align_individual_cells = F,
  align_fl_to_bf = T,
  image_type = "brightfield",
  bf_fl_mapping = "list"
)
}
\arguments{
\item{max_split_over_minor}{To-do: document or link to explanation.}

\item{max_dist_over_waist}{To-do: document or link to explanation.}

\item{max_pixels_per_cell}{To-do: document or link to explanation.}

\item{min_pixels_per_cell}{To-do: document or link to explanation.}

\item{background_reject_factor}{To-do: document or link to explanation.}

\item{tracking_comparison}{To-do: document or link to explanation.}

\item{align_individual_cells}{To-do: document or link to explanation.}

\item{align_fl_to_bf}{To-do: document or link to explanation.}

\item{image_type}{To-do: document or link to explanation.}

\item{bf_fl_mapping}{To-do: document or link to explanation.}
}
\value{
A nice list of parameters.
}
\description{
Returns a list of key-value pairs, for the default Cell-ID parameters.
}
\details{
Documentation for each parameter can be found at: https://github.com/darksideoftheshmoo/cellID-linux#parameters

Boolean values are for "flag" type parameters
which enable a feature when present (eg. "align_fl_to_bf"),
or, if absent, indicate default behavior.

Other parameters have values
which must end up separated from names by a space " "
in the parameters.txt file format that Cell-ID uses:

\preformatted{
max_split_over_minor 0.5
max_dist_over_waist 8
max_pixels_per_cell 2000
min_pixels_per_cell 75
background_reject_factor 0.75
tracking_comparison 0.2
align_fl_to_bf
image_type brightfield
bf_fl_mapping list
}
}
\seealso{
\link[rcell2]{parameters.write}, \link[rcell2]{arguments}
}
