% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellid_wrapper.R
\name{parameters.write}
\alias{parameters.write}
\title{Write parameters to a [temporary] file}
\usage{
parameters.write(
  parameters.list = rcell2::parameters.default(),
  param.dir = base::tempdir(),
  param.file = NULL
)
}
\arguments{
\item{parameters.list}{a parameters list for Cell-ID (like one from parameters.default)}

\item{param.dir}{directory where parameter files will be written.}

\item{param.file}{a file name for the parameters file.}
}
\value{
A path to the text file where parameters where written.
}
\description{
Parses a \code{parameters.list} list object, and saves its contents in a Cell-ID friendly text file.
}
\seealso{
\link[rcell2]{parameters.list}, \link[rcell2]{arguments}
}
