% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hu_moments_short.R
\name{pic_df_from_tiff}
\alias{pic_df_from_tiff}
\title{Generate a dataframe with boundary coordinates from mask tiff file}
\usage{
pic_df_from_tiff(tiff_path, image_bits, cell_id_offset = -1)
}
\description{
Generate a dataframe with boundary coordinates from mask tiff file
}
