% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definitions.R
\name{pip}
\alias{pip}
\title{Polygon filtering function using "sp" package}
\usage{
pip(
  points,
  pgn,
  points_x_column = 1,
  points_y_column = 2,
  pgn_x_column = 1,
  pgn_y_column = 2
)
}
\description{
Polygon filtering function using "sp" package
}
