% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definitions_shinycell_utils.R
\name{plot_bound_filters}
\alias{plot_bound_filters}
\title{Plot bound shinyCell polygon filters by variable pairs}
\usage{
plot_bound_filters(bound_filters)
}
\arguments{
\item{bound_filters}{The output of \code{bind_filters}.}
}
\value{
Plots for the polygons in bound_filters.
}
\description{
Useful to check out what areas the filters are covering.
}
