% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definitions_shinycell_utils.R
\name{plot_filters}
\alias{plot_filters}
\title{Plot shinyCell polygon filters}
\usage{
plot_filters(saved_data, print_plots = TRUE)
}
\arguments{
\item{saved_data}{The output of shinyCell.}

\item{print_plots}{Set to false to prevent printing the plots on execution.}
}
\value{
Prints a lis
}
\description{
Useful to check out what areas the filters are covering.
}
