% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectFromPlot.R
\name{plotApp}
\alias{plotApp}
\title{Filtrar cdata usando gráficos y dibujando regiones}
\usage{
plotApp(user_plot, debug_messages = F)
}
\arguments{
\item{user_plot}{A ggplot}

\item{debug_messages}{print stuff}
}
\value{
Lots of stuff.
}
\description{
Filtrar cdata usando gráficos y dibujando regiones
}
