% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSubsetApp_server.R
\name{plotAppServer}
\alias{plotAppServer}
\title{Shiny app server function for shinyCell
Define server logic required to draw a histogram}
\usage{
plotAppServer(input, output, session)
}
\arguments{
\item{input}{provided by shiny}

\item{output}{provided by shiny}

\item{session}{provided by shiny}
}
\description{
Shiny app server function for shinyCell
Define server logic required to draw a histogram
}
