% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definitions.R
\name{polyFilterApply}
\alias{polyFilterApply}
\title{Apply polygonal filters to cdata}
\usage{
polyFilterApply(
  polygon_df_list,
  cdata,
  truthMode = "all",
  cell_unique_id_field = "ucid"
)
}
\arguments{
\item{polygon_df_list}{A list of polygon dataframes with columns: x (values) y (values) xvar (variable name for x values) yvar (variable name for y values) type ("Subtractive" or "Additive")}

\item{cdata}{A "cdata" dataframe.}

\item{truthMode}{Logical priority for "Subtractive" and "Additive" polygon filter types, passed to \code{calculateTruth}. Should be one of "all" (Subtractive overcomes Additive) or "any" (Additive overcomes Subtractive).}

\item{cell_unique_id_field}{Name for the column holding the unique identifier (a "primary key") for each data point (i.e. the "ucid" is not suficcient for time series datasets).}
}
\value{
a "saved_data" list object, where the cdata is appended a "filter" logical column.
}
\description{
Apply polygonal filters to cdata
}
