% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definitions.R
\name{polyFilterCell}
\alias{polyFilterCell}
\title{Build filterDF from polygonDF and cdataDF}
\usage{
polyFilterCell(cdataDF, filterDF, polygonDF, polygonName = 1)
}
\description{
Build filterDF from polygonDF and cdataDF
}
