% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_filter.R
\name{qc_filter}
\alias{qc_filter}
\title{qc_filter}
\usage{
qc_filter(X, filter, subset = NULL)
}
\arguments{
\item{X}{cell.data object}

\item{filter}{conditions to filter}

\item{subset}{subset of cells to select for filtering}
}
\value{
cell.data object
}
\description{
Filters cells that DO NOT MATCH the condition in \code{filter}. i.e. if you want
to REMOVE cells with a.tot < 100, call \code{qc_filter(X, a.tot > 100)}.
In addition, subsets selects a subset of cells to which apply the filter.
i.e. to apply the same area filter but only to cells with f.tot.y values under 50,
call \code{qc_filter(X, a.tot > 100, subset = f.tot.y < 50)}
}
\details{
It does not delete the filtered rows, it only modifies the qc variable. To effectively
remove them, use \link{qc_execute}.
}
