% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_filter.R
\name{qc_reset}
\alias{qc_reset}
\title{qc_reset}
\usage{
qc_reset(X)
}
\arguments{
\item{X}{}
}
\value{
cell.data object
}
\description{
resets the \code{QC} variable to \code{TRUE}, removing all applied \code{qc_filters}.
}
