% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_tiff_tags.R
\name{read_description}
\alias{read_description}
\title{Extract additional info from a `TIFFTAG_DESCRIPTION`.}
\usage{
read_description(x)
}
\value{
A list with the `description` attribute modified.
}
\description{
Extract additional info from a `TIFFTAG_DESCRIPTION`.
}
\keyword{internal}
