% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ijtiff_read.R
\name{read_tags}
\alias{read_tags}
\alias{tags_read}
\title{Read TIFF tag information without actually reading the image array.}
\usage{
read_tags(path, frames = 1, list_safety = "error", msg = TRUE)

tags_read(path, frames = 1)
}
\arguments{
\item{path}{A string. The path to the tiff file to read.}

\item{frames}{Which frames do you want to read tags from. Default first frame
only. To read from the 2nd and 7th frames, use `frames = c(2, 7)`, to read
from all frames, use `frames = "all"`.}

\item{list_safety}{A string. This is for type safety of this function. Since
returning a list is unlikely and probably unexpected, the default is to
error. You can instead opt to throw a warning (`list_safety = "warning"`)
or to just return the list quietly (`list_safety = "none"`).}

\item{msg}{Print an informative message about the image being read?}
}
\value{
A list of lists.
}
\description{
TIFF files contain metadata about images in their _TIFF tags_. This function
is for reading this information without reading the actual image.
}
\examples{
read_tags(system.file("img", "Rlogo.tif", package = "ijtiff"))
read_tags(system.file("img", "Rlogo-banana.tif", package = "ijtiff"),
  frames = c(2, 4)
)
}
\seealso{
[read_tif()]
}
\author{
Simon Urbanek, Kent Johnson, Rory Nolan.
}
