% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cell_mask_data.R
\name{read_tiff_masks}
\alias{read_tiff_masks}
\title{Read cellID masks from BF.out TIFF image}
\usage{
read_tiff_masks(path, cell_id_offset = -1)
}
\arguments{
\item{path}{A string. The path to the 16-bit BF.out tiff file to read.}

\item{cell_id_offset}{the offset with respect to maximum pixel intensity, such
that \code{cellID = maximum_intensity - boundary_intensity + cell_id_offset}.}
}
\value{
A data frame.
}
\description{
Reads boundary and interior pixel data sorted by cellID from a BF.out TIFF 
image. Expects a 16-bit image with masks encoded on a blank background.
}
