% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_tiff_tags.R
\name{read_tiff_tags}
\alias{read_tiff_tags}
\title{Read TIFF tag information.}
\usage{
read_tiff_tags(path, frames = 1)
}
\arguments{
\item{path}{A string. The path to the tiff file to read.}

\item{frames}{Which frames do you want to read tags from. Default first frame
  only. To read from the 2nd and 7th frames, use `frames = c(2, 7)`, to read
  from all frames, use `frames = "all"`.
r}
}
\value{
A list of lists.
}
\description{
TIFF files contain metadata about images in their \emph{TIFF tags}. This function
is for reading this information without reading the actual image. It extends
on the \code{read_tags} function from the \code{ijtiff} package, and extracts
additional information nested within the \code{description} attribute of some TIFF
images.
}
\examples{
rcell2::read_tiff_tags(path = "data/image_samples/BF_Position001.tif", frames = 1)

}
