% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellid_wrapper.R
\name{read_tsv.con.pos}
\alias{read_tsv.con.pos}
\title{Una función que lea un .csv y les agregue una columna con un id del archivo (pos)}
\usage{
read_tsv.con.pos(.nombre.archivo, .carpeta, position.pattern, col_types = "c")
}
\description{
Una función que lea un .csv y les agregue una columna con un id del archivo (pos)
}
