% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyCell.R
\name{shinyCell}
\alias{shinyCell}
\title{Filtrar cdata usando gráficos y dibujando regiones}
\usage{
shinyCell(
  cdata,
  pdata = NULL,
  paths,
  filters = list(),
  filters.init_selected = T,
  plotType = "Dots",
  seed = 1,
  initial_facet = "",
  initial_vars = c("a.tot", "fft.stat"),
  facet_grid_option = TRUE,
  facets_scale_free = "fixed",
  n_max = 100,
  boxSize = 80,
  filter_progress_file = NULL,
  launch.browser = F,
  skip_input_check = F,
  ...
)
}
\arguments{
\item{cdata}{A Rcell "cdata" data.frame with the CellID variables (not the cell.data object, \code{cell.data$data}).}

\item{pdata}{An optional "pdata" data.frame, with positions' metadata (NULL by default).}

\item{paths}{A "paths" data.frame, with paths to each positions' images (i.e. \code{cell.data$images}).}

\item{filters}{An optional list with the filters from a previous shinyCell run (dataframes with points of 2D polygons). An empty \code{list()} by default.}

\item{plotType}{Type of the filtering plot, either: "Dots" (point scatterplot, defaut), "Hex" (2D histogram with hexagonal bins), "Density", "Pics" (a \code{cellSpread} plot).}

\item{seed}{Seed value for sampling of cell images.}

\item{initial_facet}{Initial ggplot facet formula as a string (for example: "~pos+group_1")}

\item{initial_vars}{Initial cdata variables as a character vector (defaults to \code{c('a.tot', 'fft.stat')}).}

\item{facet_grid_option}{Use ggplot's facet_grid (TRUE, default) or facet_wrap (FALSE).}

\item{facets_scale_free}{Use ggplot's facets with fixed scales (NULL, default) or free scales ("free").}

\item{boxSize}{Size in pixels for individual cells' images.}

\item{filter_progress_file}{Path to an RDS file, used for saving filtering progress (in case something goes wrong). Using FALSE disables this feature. Set to NULL (the default) to let tempfile() choose a path for the RDS, or set to a valid path of your choice.}

\item{launch.browser}{Set to \code{'firefox'} or equivalent to launch the app in-browser (\code{FALSE} by default). Useful when launching fails with error \code{Error in utils::browseURL(appUrl)} or similar.}

\item{skip_input_check}{If FALSE (default)}

\item{...}{Further arguments passed to \code{magickCell()}.}
}
\value{
A named list with the original cdata and a list of filters. The cdata includes an extra "filter" column, indicating if a row is to be kept (TRUE) or filtered out (FALSE). The list of filters can be passed as a filter argument, and can be plotted with \code{plot_filters}.
}
\description{
Filtrar cdata usando gráficos y dibujando regiones
}
\examples{

# Minimal example:

path <- "/path/to_your/cellid_images/"

cell.data <- rcell2::cell.load.alt(path = path)

cdata <- cell.data$data  # CellID dataframe

images <- cell.data$images  # Image paths

pdata <- read.csv("data/pdata.csv")  # "Position" metadata

filter.output <- 
  rcell2::shinyCell(cdata = cdata, 
                    pdata = pdata, 
                    paths = images)
                    
plot_filter(filter.output)

cdata.filtered <- dplyr::filter(filter.output, filter)
  
}
