% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spline.poly.R
\name{smooth.spline.poly}
\alias{smooth.spline.poly}
\title{whuber's spline.poly adapted to stats::smooth.spline}
\usage{
smooth.spline.poly(xy, k = 3, dof = 5, ...)
}
\arguments{
\item{xy}{A data.frame with ordered X/Y pairs representing the polygon}

\item{k}{The amount of points used to "close" the path, and prevent discontinuous derivatives at the end-points.}

\item{dof}{The degrees of freedom used by \code{stats::smooth.spline}}
}
\value{
Returns a data.frame of points.
}
\description{
See: https://gis.stackexchange.com/a/24929
}
\details{
Splining a polygon with \code{stats::smooth.spline}

  The rows of 'xy' give coordinates of the boundary vertices, in order.
  'vertices' is the number of spline vertices to create.
             (Not all are used: some are clipped from the ends.)
  'k' is the number of points to wrap around the ends to obtain
      a smooth periodic spline.
}
