% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spline.poly.R
\name{spline.poly}
\alias{spline.poly}
\title{whuber's spline.poly as shared in StackExchange}
\usage{
spline.poly(xy, vertices, k = 3, ...)
}
\arguments{
\item{xy}{A data.frame with ordered X/Y pairs representing the polygon}

\item{vertices}{The amount of vertices in the final smoothing/interpolation.}

\item{k}{The amount of points used to "close" the path, and prevent discontinuous derivatives at the end-points.}
}
\value{
Returns an array of points.
}
\description{
See: https://gis.stackexchange.com/a/24929
}
\details{
Splining a polygon with \code{stats::spline}

  The rows of 'xy' give coordinates of the boundary vertices, in order.
  'vertices' is the number of spline vertices to create.
             (Not all are used: some are clipped from the ends.)
  'k' is the number of points to wrap around the ends to obtain
      a smooth periodic spline.
}
