% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k-means_clustering.R
\name{standardise_data}
\alias{standardise_data}
\title{standardise_data}
\usage{
standardise_data(x)
}
\arguments{
\item{x}{a data.frame with numeric variables suitable for normalization}
}
\value{
a data.frame with normalized data
}
\description{
Standardise data column-wise by calculating the respective Z-scores, i.e., subtracting the mean and dividing by the standard deviation.
}
