% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k-means_functions.R
\name{standardize_data}
\alias{standardize_data}
\title{standardize_data}
\usage{
standardize_data(x)
}
\arguments{
\item{x}{a data.frame with numeric variables suitable for normalization}
}
\value{
a data.frame with normalized data
}
\description{
Standardize data column-wise by calculating the respective Z-scores, i.e., subtracting the mean and dividing by the standard deviation.
}
