% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagCell.R
\name{tagCell}
\alias{tagCell}
\title{Filtrar cdata usando gráficos y dibujando regiones}
\usage{
tagCell(
  cdata,
  pdata,
  paths,
  cell_tags,
  filters = list(),
  tag_box_size = 50,
  n_max = 10,
  seed = 1,
  tmp_csv_output = tempfile(tmpdir = "./", fileext = ".txt"),
  tag_ggplot = NULL,
  ...
)
}
\arguments{
\item{cdata}{dataframe of "cell data"}

\item{pdata}{dataframe "position data"}

\item{paths}{dataframe of image paths}

\item{cell_tags}{list of named vectors corresponding to tag groups and tags.}

\item{tag_box_size}{size of the image crop in pixels}

\item{n_max}{max number of boxes in the image}

\item{seed}{seed for random sampling of images}

\item{tmp_csv_output}{file path into which tagging information will be dumped progressively}

\item{tag_ggplot}{a ggplot object to display in the second tab, may be used for something someday.}
}
\value{
Lots of stuff.
}
\description{
Ejemplo:
saved <- rcell2::tagCell(cell.data$d, pdata, paths, 
                         cell_tags = list(far1_drop = c(TRUE,
                                                        FALSE),
                                          artifact = c("segmentation",
                                                       "death")
                         ),  # list(named_item1 = c(option1, option2, ...), named_item2 ...)
                         tag_box_size = 75,
                         tag_ggplot = p)
}
