% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagCell_server.R
\name{tagCellServer}
\alias{tagCellServer}
\title{Shiny app server function for tagCell
Define server logic required to draw a histogram}
\usage{
tagCellServer(input, output, session)
}
\arguments{
\item{input}{provided by shiny}

\item{output}{provided by shiny}

\item{session}{provided by shiny}
}
\description{
Shiny app server function for tagCell
Define server logic required to draw a histogram
}
