% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagCell.R
\name{tags.to.onehot}
\alias{tags.to.onehot}
\title{Pivot cell tags to a cdata-joinable dataframe, with one hot encoding}
\usage{
tags.to.onehot(tags.df, exclude.cols = c("pos", "cellID", "viewed"))
}
\arguments{
\item{tags.df}{Output from tagCell.}

\item{exclude.cols}{Character vector with names of columns which should be removed from input.}
}
\description{
Pivot cell tags to a cdata-joinable dataframe, with one hot encoding
}
\details{
Ver:
* `~/Projects/Academia/Doctorado/gitlabs_acl/rtcc/far1/analisis_Far1_arresto-lavado/R/analisis_pos_2_a_7_v7_tags_analysis.Rmd`
* https://stackoverflow.com/questions/55288338/r-multi-hot-encoding-among-multiple-columns
* https://stackoverflow.com/a/63454411/11524079
}
