% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_mask_data.R
\name{tsv_paths_from_args}
\alias{tsv_paths_from_args}
\title{Generate data.frame with paths to coordinate TSV files}
\usage{
tsv_paths_from_args(
  arguments,
  position_pattern = ".*Position(\\\\d+)$",
  tsv_pattern = "^out_all_masks.tsv$"
)
}
\arguments{
\item{arguments}{An "arguments" dataframe, as produced by \code{rcell2.cellid::arguments}.}

\item{position_pattern}{A regex pattern with one group for the integer position number, extracted from the directory name holding the TSV file.}

\item{tsv_pattern}{A regex pattern matching the name of the TSV file.}
}
\description{
Convenience function to generate input for \code{hues_from_tsv_files2}.
}
