% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{update_ntot}
\alias{update_ntot}
\title{Update number of frames a cell appears in}
\usage{
update_ntot(object)
}
\arguments{
\item{object}{cell.data}
}
\value{
cell.data
}
\description{
Update number of frames a cell appears in
}
