% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_cell_data.R
\name{write.delim}
\alias{write.delim}
\title{write tab delimited file}
\usage{
write.delim(x, file = "", quote = FALSE, sep = "\\t", row.names = FALSE, ...)
}
\arguments{
\item{x}{data frame to write}

\item{file}{file to output}

\item{quote}{a logical value (TRUE or FALSE) or a numeric vector. If TRUE, any character
or factor columns will be surrounded by double quotes. If a numeric vector, its elements
are taken as the indices of columns to quote. In both cases, row and column names are
quoted if they are written. If FALSE, nothing is quoted.}

\item{sep}{the field separator string. Values within each row of x are separated by this string.}

\item{row.names}{either a logical value indicating whether the row names of x are to be written
along with x, or a character vector of row names to be written.}

\item{...}{arguments to write.table: append, col.names, sep, dec and qmethod cannot be altered.}
}
\description{
Writes a tab delimited file. Wrapper to write.table
}
