% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hu_moments_short.R
\name{append_hues2}
\alias{append_hues2}
\title{Add Hu moments data.frame to a cell_data object}
\usage{
append_hues2(
  tsv_files_df,
  cell_data = list(),
  return_points = F,
  parralellize = T,
  shape_pixtype = "b",
  shape_flagtype = 0,
  overwrite = F
)
}
\arguments{
\item{cell_data}{An empty list by default. You may pass a cell.data list object from Rcell2 load_cell_data for convenience.}

\item{return_points}{if TRUE it will add a "masks" dataframe to the cell_data object, containing the mask coordinates.}

\item{shape_pixtype}{Default "b" for Hu moments based on boundary points. A character vector containing any of c("b", "i").}

\item{shape_flagtype}{Default 0 for Hu moments based on flag value 0. Can be any of the integer flag values present in the \code{out_bf_fl_mapping} CellID files.}

\item{overwrite}{Default FALSE, overwrite Hu moments element in the input list if already present.}

\item{tsv_files}{Paths to the XY coordinates TSV files from CellID mask_mod "-t" option.}
}
\value{
The Hu moments dataframe usis assgned to the input list as an element named "Hu_moments".
}
\description{
The required TSV are generated by the mask_mod branch of the CellID program found at: https://github.com/darksideoftheshmoo/cellID-linux/tree/mask_mod
}
\details{
To play with the XY coordinates of the TIFF masks you may want to set \code{return_points = T}.
}
