% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellid_wrapper.R
\name{arguments_to_images}
\alias{arguments_to_images}
\title{Make and "images" dataframe from "arguments" dataframe}
\usage{
arguments_to_images(arguments)
}
\arguments{
\item{arguments}{The "arguments" dataframe, output from \code{rcell2::arguments()}.}
}
\value{
A data.frame similar to \code{cell.load.alt()$images}.
}
\description{
The images dataframe is needed by many rcell2 functions. If it is not available from the output of \code{load_cell_data} or \code{cell.load.alt}, then this function can help.
}
\details{
It essentially does a pivot_longer of the arguments.
}
