% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellid_wrapper.R
\name{cell.images.out}
\alias{cell.images.out}
\title{Filtrar cdata usando gráficos y dibujando regiones}
\usage{
cell.images.out(
  path,
  BF.pattern = "BF_Position\\\\d+\\\\.tif$",
  O.pattern = ".*(Position\\\\d+).*\\\\.tif",
  out.dir = "out"
)
}
\arguments{
\item{path}{directory where images are stored, full path.}

\item{BF.pattern}{regex pattern to ?FP images}

\item{O.pattern}{regex pattern to output directories}

\item{out.dir}{name for output directories paths "out"}
}
\value{
A vector of directories.
}
\description{
Filtrar cdata usando gráficos y dibujando regiones
}
