% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellid_wrapper.R
\name{cell.load.alt}
\alias{cell.load.alt}
\title{Cargar el output de cell-id}
\usage{
cell.load.alt(
  path = "data/images2/",
  pdata = NULL,
  position.pattern = "Position\\\\d+",
  cellid.zero.pad = 4,
  ...
)
}
\arguments{
\item{path}{Path to images directory}

\item{position.pattern}{Regex describing what the position string looks like (default 'Position\\d+')}

\item{...}{pass "pdata" with the path to the pdata file to merge it internally}

\item{ucid.zero.pad}{amount of decimal digits for the cellID (defaults 4, corresponding to a maximum of 9.999 cellIDs and 9999 positions)}
}
\value{
a data frame
}
\description{
Cargar el output de cell-id
}
