% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellid_wrapper.R
\name{cell.load.alt}
\alias{cell.load.alt}
\title{Cargar el output de cell-id}
\usage{
cell.load.alt(
  path,
  pdata = NULL,
  position.pattern = ".*Position(\\\\d+).*",
  ucid.zero.pad = 4,
  ...
)
}
\arguments{
\item{path}{Path to images directory}

\item{pdata}{Path to metadata CSV file}

\item{position.pattern}{Regex describing what the position string looks like (default ".*Position(\\d+).*") including a capturing group for the position ID number (coerced to integer).}

\item{ucid.zero.pad}{amount of decimal digits for the cellID (defaults 4, corresponding to a maximum of 9.999 cellIDs and 9999 positions)}

\item{...}{Arguments passed on to \code{cargar.out_all}. Patterns for "out" files, fluorescence channel, and other options may be changed here.}
}
\value{
A list of dataframes: data (CellID data), images (images metadata and paths), image_maping (extra mapping metadata from CellID: BF to FL correspondence, channel flag, bf_as_fl flag, and one-letter channel encoding).
}
\description{
Cargar el output de cell-id
}
