% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cell_mask_data.R
\name{cell.load.boundaries}
\alias{cell.load.boundaries}
\title{Load cellid boundary data to a dataframe}
\usage{
cell.load.boundaries(
  data.source,
  ...,
  arguments = NULL,
  pixel.type = "b",
  close.paths = FALSE,
  cell.data = NULL,
  tiff.channel = "BF.out"
)
}
\arguments{
\item{data.source}{Either "out.tif" or "masks.tsv"}

\item{arguments}{The arguments dataframe used to run cellid (prepared with \code{rcell2::arguments}).}

\item{pixel.type}{When \code{data.source = "masks.tsv"}, you may choose the pixel "type". At least one of \code{c("i", "b")} for interior and/or boundary pixels ("b" by default).}

\item{close.paths}{When TRUE and \code{data.source = "masks.tsv"}, append the first row to the end of the data.frame (groping by cellID and pos). Useful for plotting of making closed polygons.}

\item{cell.data}{When \code{data.source = "out.tif"}, provide the cell.data object (prepared with \code{rcell2::cell.load.alt}).}

\item{tiff.channel}{When \code{data.source = "out.tif"}, provide the channel name for images holding boundary data ('BF.out' by default).}
}
\value{
A "cell.boundaries" data.frame.
}
\description{
Using `data_source = "out.tif"` will give **un**ordered boundary pixels.
}
\details{
Using `data_source = "masks.tsv"` will give ordered boundary pixels.
}
