% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellid_wrapper.R
\name{cell}
\alias{cell}
\title{Filtrar cdata usando gráficos y dibujando regiones}
\usage{
cell(
  cell.args,
  position.pattern = "Position\\\\d+",
  position.time.pattern = NULL,
  position.time.pattern.sep = "_?",
  fluorescence.pattern = "^(.FP)",
  cell.command = "cellBUILTIN",
  channels = c("b", "f"),
  old_dirs_path = NULL,
  old_dirs_pattern = "^Position\\\\d\\\\d\\\\d$",
  no_cores = NULL,
  label_cells = 1,
  bf_out_null_bg = 1,
  dry = F
)
}
\arguments{
\item{cell.args}{An argument list, as built by cellArgs().}

\item{position.pattern}{a regular expression that recognizes the position in the image name, such as: "Position\\d+" for Position01, Position100, etc.}

\item{position.time.pattern}{a regular expression that recognizes the time in the image name, such as: "time\\d+" for time01, time007, etc.}

\item{position.time.pattern.sep}{a regular expression that matches the pos-time separator characters (default is an optional underscore), such as: "_" for Position01_time02}

\item{fluorescence.pattern}{a regular expression that recognizes the fluorescence tag in the image name, such as "^(.FP)" for TFP, YFP, etc.}

\item{cell.command}{the CellID command, either "cellBUILTIN" for the builtin binary, a path to the binary executable (get if from https://github.com/naikymen/cellID-linux).}

\item{channels}{Default c("b", "f"), don't change, used to create temporal file lists for BF and ?FP and pass cell id arguments.}

\item{no_cores}{Position-wise parallelization,internally capped to number of positions in cell.args.}

\item{label_cells}{Set to 0 to disable labeling cells with their number.}

\item{bf_out_null_bg}{Set to 1 blank the BF out background.}

\item{dry}{Do everything without actually running CellID.}
}
\value{
Nothing.
}
\description{
Filtrar cdata usando gráficos y dibujando regiones
}
