% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellid_wrapper.R
\name{cell2}
\alias{cell2}
\title{Function to run CellID}
\usage{
cell2(
  arguments,
  cell.command = NULL,
  no_cores = NULL,
  debug_flag = 0,
  dry = F,
  label_cells_in_bf = F,
  fill_interior_pixels = F,
  output_coords_to_tsv = F,
  encode_cellID_in_pixels = F,
  ignore.stdout = T,
  intern = T
)
}
\arguments{
\item{arguments}{An argument data.frame, as built by rcell2::cellArgs2.}

\item{cell.command}{Path to the binary executable (get if from https://github.com/darksideoftheshmoo/cellID-linux). Or "cellBUILTIN" for the builtin binary.}

\item{no_cores}{Position-wise parallelization,internally capped to number of positions in cell.args.}

\item{debug_flag}{Set to 0 to disable CellID printf messages (builtin CellID only).}

\item{dry}{Do everything without actually running CellID, print the commands that would have been issued.}

\item{label_cells_in_bf}{Set to TRUE to enable labeling of cells with their CellID in the BF output image (CellID option '-l').}

\item{fill_interior_pixels}{Set to TRUE to fill each cell interior area in the output image file with intensity-labeled pixels (CellID option '-i').}

\item{output_coords_to_tsv}{Set to TRUE to write cell interior and boundary pixels data to a .tsv file in the output directory (CellID option '-m').}

\item{encode_cellID_in_pixels}{Set to TRUE to write cell interior and boundary pixels with intensity-encoded CellIDs and blank the rest of the image (CellID option '-s').}

\item{ignore.stdout}{Set to FALSE to see CellID output from a system call.}

\item{intern}{Set to TRUE to save CellID output from a system call to a file in the "out" directories (one per position) and the commands to a file at the first "path" in the arguments data.frame.}
}
\value{
A dataframe with one column indicating the issued commands. Use rcell2::load_cell_data to get the results from the CellID output, typically located at the images path.
}
\description{
Function to run CellID
}
