% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellid_wrapper.R
\name{cellArgs.print}
\alias{cellArgs.print}
\title{Print cell.args}
\usage{
cellArgs.print(cell.args, which.args = c("p", "b", "f", "o"))
}
\arguments{
\item{cell.args}{cell.args list.}

\item{which.args}{which arguments should be printed}
}
\value{
Nothing.
}
\description{
Print cell.args
}
