% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellid_wrapper.R
\name{cellArgs}
\alias{cellArgs}
\title{Filtrar cdata usando gráficos y dibujando regiones}
\usage{
cellArgs(
  path,
  parameters,
  BF.pattern = "BF_Position\\\\d+\\\\.tif$",
  FP.pattern = "FP_Position\\\\d*\\\\.tif$",
  O.pattern = ".*(Position\\\\d+)\\\\.tif",
  out.dir = "out"
)
}
\arguments{
\item{path}{directory where images are stored, full path.}

\item{parameters}{absolute path to the parameters file.}

\item{BF.pattern}{regex pattern to BF images}

\item{FP.pattern}{regex pattern to ?FP images}

\item{O.pattern}{regex pattern to output directories}

\item{out.dir}{name for output directories paths "out"}
}
\value{
Nothing.
}
\description{
Filtrar cdata usando gráficos y dibujando regiones
}
