% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definitions_magick.R
\name{cellGif}
\alias{cellGif}
\title{cellMagick configured for prodicing a ucid's gif}
\usage{
cellGif(
  cdata,
  paths,
  equalize_images = F,
  normalize_images = F,
  channels = c("BF.out"),
  animation_delay = 1/3,
  id_column = "ucid",
  time_colum = "t.frame",
  stack_channels_horizontally = TRUE,
  ...
)
}
\arguments{
\item{cdata}{A Rcell data.frame (not the object).}

\item{paths}{A paths dataframe with file path, t.frame, position and channel information of each picture.}

\item{equalize_images}{Use magick's function to "equalize" the image when TRUE (FALSE by default).}

\item{normalize_images}{Use magick's function to "normalize" the image when TRUE (FALSE by default).}

\item{channels}{Name of the CellID channel (BF, BF.out, RFP, etc.). "BF.out" by default.}

\item{animation_delay}{Delay between animation frames in seconds.}

\item{stack_channels_horizontally}{Time increases from left to right (TRUE) or from up to down (FALSE).}

\item{...}{Arguments passed on to magickCell.}
}
\description{
cellMagick configured for prodicing a ucid's gif
}
