% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellid_wrapper.R
\name{cellid}
\alias{cellid}
\title{Correr cellid desde C}
\usage{
cellid(args, label_cells = 1)
}
\arguments{
\item{args}{el comando de cellid entero, tal como se ejecutaria en bash "cell -p ..."}

\item{label_cells}{Set to 0 to disable labeling cells with their number.}
}
\value{
Nada, el output está en los directorios.
}
\description{
Correr cellid desde C
}
