% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellid_wrapper.R
\name{cellid}
\alias{cellid}
\title{Correr Cell-ID desde R usando .C()}
\usage{
cellid(args, debug_flag = 0)
}
\arguments{
\item{args}{el comando de cellid entero, tal como se ejecutaria en bash "cell -p ..."}

\item{debug_flag}{Set to 0 to disable CellID printf messages.}
}
\value{
Exit code from CellID
}
\description{
Correr Cell-ID desde R usando .C()
}
