% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definitions_magick.R
\name{cellSpread}
\alias{cellSpread}
\title{2D binning of data and tiling of cell pictures}
\usage{
cellSpread(
  cdata,
  paths,
  ch = "BF.out",
  boxSize = 80,
  xvar = "a.tot",
  yvar = "fft.stat",
  x.cuts = 7,
  y.cuts = 7,
  for_plotting = F,
  ...
)
}
\arguments{
\item{cdata}{A Rcell data.frame (not the object).}

\item{paths}{A paths dataframe with file path, t.frame, position and channel information of each picture.}

\item{ch}{Name of the CellID channel (BF, BF.out, RFP, etc.). "BF.out" by default.}

\item{boxSize}{Size of the box containing the individual cell images. 50 by default.}

\item{xvar, yvar}{Strings indicating names for the variables to plot in the horizontal (x) and vertical (y) axis.}

\item{x.cuts, y.cuts}{Integers indicating the number of cuts for each variable.}

\item{for_plotting}{Return value changes to list of elements important for plotting.}

\item{...}{Arguments passed on to magickCell.}
}
\description{
2D binning of data and tiling of cell pictures
}
