% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definitions_magick.R
\name{cellStrips}
\alias{cellStrips}
\title{Wraps cellMagick to make strips, optionally cutting them.}
\usage{
cellStrips(
  cdata,
  paths,
  n_ucids = NULL,
  cut_breaks = 1,
  split_col = "ucid",
  ch = c("BF.out", "YFP.out"),
  sortVar = "t.frame",
  ...
)
}
\arguments{
\item{cdata}{A Rcell data.frame (not the object).}

\item{paths}{A paths dataframe with file path, t.frame, position and channel information of each picture.}

\item{n_ucids}{will select the first `n_ucids`}

\item{...}{Arguments passed on to magickCell.}

\item{cut_strips}{Use `cut` to split the image series.}
}
\description{
`cdata` is split by `split_col` and then images are generated.
}
\details{
`images` are split with `cut`, useful wen strips are too long.
}
