% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hu_moments_short.R
\name{check_tiff_mask}
\alias{check_tiff_mask}
\title{Make plots to examine correspondence between mean mask xpos/ypos and CellID xpos/ypos for each cell.}
\usage{
check_tiff_mask(cell_data)
}
\description{
Make plots to examine correspondence between mean mask xpos/ypos and CellID xpos/ypos for each cell.
}
