% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cimage.R
\name{get_cell_image}
\alias{get_cell_image}
\title{Title}
\usage{
get_cell_image(X, ...)
}
\arguments{
\item{X}{cell.data object or data.frame that specifies the images}

\item{...}{further arguments for methods}
}
\value{

}
\description{
Title
}
