% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k-means_functions.R
\name{get_fit_vars}
\alias{get_fit_vars}
\title{get_fit_vars}
\usage{
get_fit_vars(x, f.channels, var_cats = NULL, custom_vars = NULL)
}
\arguments{
\item{x}{a cell.data data.frame}

\item{var_cats}{a string value indicating whether morphology ('m') and/or fluorescence ('f') variables should be used for clustering.}
}
\value{
a reduced cell.data data.frame with selected clustering variables only
}
\description{
Extract specific variables relevant for clustering
}
