% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worflow_helper.R
\name{get_workflow_template}
\alias{get_workflow_template}
\title{A function to donwload the latest worflow tempalte in Rmarkdown}
\usage{
get_workflow_template(file_name = "rcell2_workflow_template.Rmd")
}
\arguments{
\item{file_name}{File name for the wokflow template.}
}
\value{

}
\description{
Will donwload the .Rmd file to the current working directory.
}
