% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definitions.R
\name{getFacetVars}
\alias{getFacetVars}
\title{Una función para procesar el facet string y que devuelva las variables presentes en names(pdata)}
\usage{
getFacetVars(pdata, facetFormulaString = "pos ~ treatment")
}
\description{
Una función para procesar el facet string y que devuelva las variables presentes en names(pdata)
}
