% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hu_moments_short.R
\name{hues_from_xy2}
\alias{hues_from_xy2}
\title{Generate Hu moments from XY coordinates dataframe}
\usage{
hues_from_xy2(pic_df, split_col = "cellID")
}
\description{
This function requires a CellID column to split the coordinates by.
}
