% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagej_macros.R
\name{imagej.fft.filter}
\alias{imagej.fft.filter}
\title{Run ImageJ FFT filter macro from R}
\usage{
imagej.fft.filter(
  pic.path,
  imagej.path = "~/Software/Fiji.app/ImageJ-linux64",
  script.path = system.file("inst/FFT_filter_on_BFs_R.txt", package = "rcell2")
)
}
\arguments{
\item{pic.path}{Path to the directory containing the image files.}

\item{imagej.path}{Path to the ImageJ binary (a path to "ImageJ-linux64" or equivalent).}

\item{script.path}{Path to the ImageJ macro. Defaults to built-in macro.}
}
\description{
Replaces files matching \code{BF.*.tif} in the target directory.
}
