% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definitions_magick.R
\name{magickCell}
\alias{magickCell}
\title{Funcion copada para mostrar fotos basada en magick}
\usage{
magickCell(
  cdata,
  paths,
  max_size = 500,
  cell_resize = 100,
  boxSize = 50,
  n = 100,
  ch = "BF.out",
  sortVar = "xpos",
  seed = 1,
  .debug = FALSE
)
}
\arguments{
\item{cdata}{A Rcell data.frame (not the object).}

\item{paths}{Paths a la imagen de cada posición.}

\item{max_size}{Maximum size of the final image (this resize is applied last) in pixels. 500 by default.}

\item{cell_resize}{Size of the individual cell images. "100x100" by default.}

\item{boxSize}{Size of the box containing the individual cell images. 50 by default.}

\item{n}{maximum number of cells to display.}

\item{ch}{Name of the CellID channel (BF, BF.out, RFP, ...). "BF.out" by default.}

\item{sortVar}{Variable used to sort the cells. "xpos" by default.}

\item{seed}{Seed value for sampling of cell images.}
}
\value{
Lots of stuff.
}
\description{
Funcion copada para mostrar fotos basada en magick
}
