% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magick_helpers.R
\name{magickForKnitr}
\alias{magickForKnitr}
\title{Display an image in rmarkdown with knitr}
\usage{
magickForKnitr(
  imgs,
  .prefix = "tile",
  .resize = NULL,
  .path = tempdir(),
  .include = F,
  .print = T
)
}
\arguments{
\item{imgs}{a magick image or cellMagick output}

\item{.prefix}{A string prepended to the file name, "" by default.}

\item{.resize}{a cellMagick image resize string as "200x200" (default NULL, for no resizing).}

\item{.path}{Directory where the output should be saved.}

\item{.include}{Set to true to use knitr::include_graphics directly.}

\item{.print}{Print the image path.}
}
\value{
An path to a temporary image file.
}
\description{
Display an image in rmarkdown with knitr
}
