% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magick_helpers.R
\name{magickForKnitr}
\alias{magickForKnitr}
\title{Display an image in rmarkdown with knitr}
\usage{
magickForKnitr(imgs, .resize = "200x200")
}
\arguments{
\item{imgs}{a cellmagick image}

\item{.resize}{a cellmagick image resize string (default "200x200").}
}
\value{
An path to a temporary image file.
}
\description{
Display an image in rmarkdown with knitr
}
