% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definitions_hex.R
\name{myhexbin}
\alias{myhexbin}
\title{Funci칩n hexbin
Esencialmente hexbin() pero un poco m치s a mi gusto}
\usage{
myhexbin(bindata, varx, vary, xbins = 25)
}
\description{
Funci칩n hexbin
Esencialmente hexbin() pero un poco m치s a mi gusto
}
