% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellid_wrapper.R
\name{\%>\%}
\alias{\%>\%}
\title{Pipe}
\arguments{
\item{lhs, rhs}{specify what lhs and rhs are}
}
\description{
This is not a pipe
}
\examples{
# some examples if you want to highlight the usage in the package
}
