% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSubsetApp.R
\name{plotApp}
\alias{plotApp}
\title{Filtrar cdata usando gráficos y dibujando regiones}
\usage{
plotApp(user_plot, debug_messages = F, print_plot_on_exit = F)
}
\arguments{
\item{user_plot}{A ggplot object, ready to render.}

\item{debug_messages}{Print internal stuff, useful for debugging.}

\item{print_plot_on_exit}{Prints user_plot replacing it's data with the "brushed" subset}
}
\value{
A data.frame with the brushed points from the original plot's data.
}
\description{
Filtrar cdata usando gráficos y dibujando regiones
}
