% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definitions.R
\name{polyFilterApply}
\alias{polyFilterApply}
\title{Apply string filters to cdata}
\usage{
polyFilterApply(
  polygon_df_list,
  cdata,
  truthMode = "all",
  cell_unique_id_field = "ucid"
)
}
\description{
Apply string filters to cdata
}
