% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definitions.R
\name{rangeExpand}
\alias{rangeExpand}
\title{range expand https://www.rosettacode.org/wiki/Range_expansion#R}
\usage{
rangeExpand(text = "1:2, 7-9", maxPos)
}
\description{
range expand https://www.rosettacode.org/wiki/Range_expansion#R
}
