% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cell_mask_data.R
\name{read_tiff_masks}
\alias{read_tiff_masks}
\title{Read cell masks from BF.out}
\usage{
read_tiff_masks(
  path,
  cell_id_offset = -1,
  interior_offset = NULL,
  blank_bg = NULL
)
}
\arguments{
\item{path}{A string. The path to the 16-bit BF.out tiff file to read.}

\item{cell_id_offset}{the offset with respect to maximum pixel intensity, such
that \code{cellID = maximum_intensity - boundary_intensity + cell_id_offset}.}

\item{interior_offset}{logical. If \code{TRUE} the function expects that
cell masks have interior pixels offset from boundary pixels.}

\item{blank_bg}{logical. If \code{TRUE} the function assumes that the image
background (i.e., non-mask pixels) are blank.}
}
\value{
A data frame of mask pixels ordered by cellID.
}
\description{
Read boundary and interior pixel data from a BF.out TIFF image.
}
\details{
This function reads cell masks from 16-bit BF.out images onto which boundary
and interior pixels have been encoded as intensity values proportional to each
object's cellID.

By default the function accounts for the possibility that the image may have
a brightfield image background and/or interior pixel intensities offset from
boundary pixels. It then first identifies the boundary pixel range, then
checks if the image also has offset interior pixels, and finally extracts
boundary and interior pixels separately. This behavior can be overridden by
explicitly specifying the image type via the \code{blank_bg} and
\code{interior_offset} parameters. This may speed up running time.
}
