% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagCell.R
\name{tagCell}
\alias{tagCell}
\title{Filtrar cdata usando gráficos y dibujando regiones}
\usage{
tagCell(
  cdata,
  pdata,
  paths,
  cell_tags,
  filters = list(),
  tag_box_size = 50,
  n_max = 10,
  seed = 1,
  tmp_csv_output = tempfile(tmpdir = "./", fileext = ".txt"),
  tag_ggplot = NULL,
  ...
)
}
\arguments{
\item{cdata}{A Rcell data.frame (not the object).}

\item{pdata}{A "pdata" data.frame with position metadata.}

\item{paths}{Paths a la imagen de cada posición.}

\item{filters}{Vector de strings con los filtros. c() by default.}

\item{seed}{Seed value for sampling of cell images.}

\item{...}{Further arguments passed to magickCell()}

\item{plotType}{"Hex", "Density", and "Dots" are available.}

\item{initial_facet}{Initial facet formula as a string.}

\item{facet_grid_option}{Use facet_grid (TRUE, default) or facet_wrap.}

\item{facets_scale_free}{Use facets with fixed scales (NULL, default) or free scales ("free").}
}
\value{
Lots of stuff.
}
\description{
Filtrar cdata usando gráficos y dibujando regiones
}
