% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagCell.R
\name{tagCell}
\alias{tagCell}
\title{Filtrar cdata usando gráficos y dibujando regiones}
\usage{
tagCell(
  cdata,
  pdata,
  paths,
  cell_tags,
  tag_box_size = 50,
  cell_resize = 100,
  tag_channels_select = c("BF"),
  n_max = 10,
  seed = 1,
  tmp_output_file = NULL,
  tag_ggplot = NULL,
  .equalize = F,
  .normalize = T,
  debug_messages = T,
  ...
)
}
\arguments{
\item{cdata}{dataframe of "cell data"}

\item{pdata}{dataframe "position data"}

\item{paths}{dataframe of image paths}

\item{cell_tags}{list of named vectors corresponding to tag groups and tags.}

\item{tag_box_size}{size of the image crop in pixels}

\item{cell_resize}{resize of the image crop in pixels}

\item{tag_channels_select}{a vector giving names for the image channels: c("BF", "YFP.out", etc....)}

\item{n_max}{max number of boxes in the image}

\item{seed}{seed for random sampling of images}

\item{tmp_output_file}{file path into which tagging information will be dumped by user request}

\item{tag_ggplot}{a ggplot object to display in the second tab, may be used for something someday.}

\item{.equalize}{Use magick's function to "equalize" the images.}

\item{.normalize}{Use magick's function to "normalize" the images.}

\item{debug_messages}{print debug messages}
}
\value{
Lots of stuff.
}
\description{
Filtrar cdata usando gráficos y dibujando regiones
}
