% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagCell.R
\name{tagCell}
\alias{tagCell}
\title{Filtrar cdata usando gráficos y dibujando regiones}
\usage{
tagCell(
  cdata,
  pdata,
  paths,
  filters = list(),
  plotType = "Hex",
  tag_ggplot = NULL,
  seed = 1,
  initial_facet = "",
  facet_grid_option = TRUE,
  facets_scale_free = NULL,
  n_max = 10,
  tmp_csv_output = tempfile(tmpdir = "./", fileext = ".txt"),
  ...
)
}
\arguments{
\item{cdata}{A Rcell data.frame (not the object).}

\item{pdata}{A "pdata" data.frame with position metadata.}

\item{paths}{Paths a la imagen de cada posición.}

\item{filters}{Vector de strings con los filtros. c() by default.}

\item{plotType}{"Hex", "Density", and "Dots" are available.}

\item{seed}{Seed value for sampling of cell images.}

\item{initial_facet}{Initial facet formula as a string.}

\item{facet_grid_option}{Use facet_grid (TRUE, default) or facet_wrap.}

\item{facets_scale_free}{Use facets with fixed scales (NULL, default) or free scales ("free").}

\item{...}{Further arguments passed to magickCell()}
}
\value{
Lots of stuff.
}
\description{
Filtrar cdata usando gráficos y dibujando regiones
}
