% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hu_moments_short.R
\name{tsv_paths_from_args}
\alias{tsv_paths_from_args}
\title{Generate data.frame with paths to coordinate TSV files}
\usage{
tsv_paths_from_args(
  arguments,
  position_pattern = ".*Position(\\\\d+)$",
  tsv_pattern = "^out_all_masks.tsv$"
)
}
\arguments{
\item{arguments}{As produced by \code{rcell2::cellArgs2}.}

\item{position_pattern}{A regex pattern with one group for the integer position number, extracted from the directory name holding the TSV file.}

\item{tsv_pattern}{A regex pattern matching the name of the TSV file.}
}
\description{
Generate data.frame with paths to coordinate TSV files
}
